requirements.txt
aiohttp            3.10.0
openai             0.28.0
pyTelegramBotAPI   4.21.0
python-dotenv      1.0.1
requests           2.32.3